feat(): 708 update change password according to new design (#765)
* feat: edit user profile

* feat: sidenav footer popup

* feat: toggling of icon of dropdown

* test: edit my-profile and side-nav dropdown

* fix: remove changes for tab styling of setting

* feat: change password according to new design

* fix

* input updated

* fix: re-route issues when resetting password

* test: add test cases of resetPassword in sideNav

* fix: resolve comments

* fix: ui issues of dropdown icon and resolve comments

* fix: resolve comments

* fix: resolve comments

* fix: test cases

---------
