+++ app/components/SideNavFooter.tsx
@@ -8,7 +8,15 @@ import ListActionMenu from "../components/ListActionMenu"
 import { routes } from "~/constants/route.constants"
 import { useUser } from "~/utils"
 
-function Footer({ currentWorkspaceId }: { currentWorkspaceId: string }) {
+function Footer({
+  currentWorkspaceId,
+  openResetPassModel,
+  setOpenResetPassModel,
+}: {
+  currentWorkspaceId: string
+  openResetPassModel: boolean
+  setOpenResetPassModel: (e: boolean) => void
+}) {
   const { t } = useTranslation()
   const navigate = useNavigate()
   const fetcher = useFetcher()
@@ -26,7 +34,9 @@ function Footer({ currentWorkspaceId }: { currentWorkspaceId: string }) {
     {
       id: "change-password",
       menuListText: t("commonConstants.changePassword"),
-      handleItemAction: () => console.log("change password"),
+      handleItemAction: () => {
+        setOpenResetPassModel(!openResetPassModel)
+      },
     },
     {
       id: "logout",
