+++ app/constants/route.constants.ts
@@ -10,6 +10,7 @@ export const routes = {
   resultGroupTest: "/results/groupByTests",
   members: "/members",
   generalSettings: "/settings/general",
+  workspaceSetting: "/settings/workspace",
   settings: "/settings",
   forgotPassword: "/forgot-password",
   endExam: "/end-assessment",
