+++ app/routes/$workspaceId/settings.tsx
@@ -95,7 +95,8 @@ export default function Settings() {
   const { currentWorkspaceId } = useLoaderData()
 
   useEffect(() => {
-    if (location.pathname === "/settings") return navigate("/settings/general")
+    if (location.pathname === routes.settings)
+      return navigate(routes.workspaceSetting)
   }, [navigate, location.pathname])
   useEffect(() => {
     const heading = document.getElementById("settings-heading")
